:truck: move read model to prescrpition context
