+++ api/src/prescription/shared/domain/read-models/CampaignParticipationOverview.js
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 import _ from 'lodash';
 
-import { CampaignParticipationStatuses, CampaignTypes } from '../../../prescription/shared/domain/constants.js';
-import { MAX_MASTERY_RATE, MINIMUM_DELAY_IN_DAYS_BEFORE_RETRYING } from '../constants.js';
+import { MAX_MASTERY_RATE, MINIMUM_DELAY_IN_DAYS_BEFORE_RETRYING } from '../../../../shared/domain/constants.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../constants.js';
 
 const { SHARED } = CampaignParticipationStatuses;
